Move assertContext to avoid import cycle
