+++ typescript/infra/src/utils/utils.ts
@@ -7,6 +7,7 @@ import path from 'path';
 
 import { AllChains, ChainName } from '@abacus-network/sdk';
 
+import { Contexts } from '../../config/contexts';
 import { ALL_KEY_ROLES, KEY_ROLE_ENUM } from '../agents/roles';
 
 export function sleep(ms: number) {
@@ -189,4 +190,18 @@ export function assertChain(chainStr: string) {
     throw Error(`Invalid chain ${chain}`);
   }
   return chain;
+}
+
+export function assertContext(contextStr: string): Contexts {
+  const context = contextStr as Contexts;
+  if (Object.values(Contexts).includes(context)) {
+    return context;
+  }
+  throw new Error(
+    `Invalid context ${contextStr}, must be one of ${Object.values(
+      Contexts,
+    )}. ${
+      contextStr === undefined ? ' Did you specify --context <context>?' : ''
+    }`,
+  );
 }
